+++ src/web/vitest.config.ts
@@ -1,0 +1,15 @@
+import { fileURLToPath } from 'node:url'
+import { mergeConfig, defineConfig } from 'vitest/config'
+import viteConfig from './vite.config'
+
+export default mergeConfig(
+  viteConfig,
+  defineConfig({
+    test: {
+      environment: 'jsdom',
+      include: ['src/**/*.{test,spec}.{ts,tsx}'],
+      globals: true,
+      root: fileURLToPath(new URL('./', import.meta.url)),
+    },
+  }),
+)